Use `uint32_t`

+++ src/hash.c
@@ -67,7 +67,7 @@ static uint32_t hash_char_convert(SEXP x) {
 
   const char* x_utf8 = CHAR_IS_UTF8(x) ? CHAR(x) : Rf_translateCharUTF8(x);
 
-  uint64_t hash = hash_int64((int64_t) *x_utf8);
+  uint32_t hash = hash_int64((int64_t) *x_utf8);
   x_utf8++;
 
   while(*x_utf8++) {
